+++ transaction/txprivacy.go
@@ -266,6 +266,12 @@ func (tx *Tx) Init(params *TxPrivacyInitParams) error {
 		outputCoins[i] = new(privacy.OutputCoin)
 		outputCoins[i].CoinDetails = new(privacy.Coin)
 		outputCoins[i].CoinDetails.SetValue(pInfo.Amount)
+		if len(pInfo.Message) > 0 {
+			if len(pInfo.Message) > privacy.MaxSizeInfoCoin {
+				return NewTransactionErr(ExceedSizeInfoOutCoinError, nil)
+			}
+		}
+		outputCoins[i].CoinDetails.SetInfo(pInfo.Message)
 
 		PK, err := new(privacy.Point).FromBytesS(pInfo.PaymentAddress.Pk)
 		if err != nil {
